Move dataset specific config into the same files as tasks
