--- heareval/tasks/config/__init__.py
@@ -1,3 +1,0 @@
-from .dataset_config import DatasetConfig, PartitionedDatasetConfig
-from .speech_commands import SpeechCommandsConfig
-from .nsynth_pitch import NSynthPitchConfig
